Add files via upload
Initiate Parse.java to start the process.
Parse.java parses the xml and makes urls. It then puts them in an array alongside ArticleTitles and sends them to http.java
http.java sends requests to the urls using a for loop and parses them for the PMID, which is then added to an array and sent to outputXML.java
outputXML.java just puts the titles and pmids together into an xml.

There is an issue with the api-key which caused on entry to be messed up.
There is another issue with the parser that caused four of the entries to be messed up.